feat: swagger format setup
