+++ app.js
@@ -5,13 +5,50 @@ if (process.env.NODE_ENV !== 'production') {
   require('dotenv').config()
 }
 
+
 require('./models')
+
+const swaggerJsDoc = require('swagger-jsdoc')
+const swaggerUi = require('swagger-ui-express')
+
+const swaggerOptions = {
+  swaggerDefinition: {
+    openapi: '3.0.0',
+    info: {
+      title: "Simple Twitter API",
+      version: '1.0.0',
+      description: "Simple Twitter API documentation",
+      contact: {
+        name: 'support',
+        email: 'support@example.com'
+      }
+    },
+    servers: [
+      {
+        url: 'http://localhost:3000',
+        description: 'Development server'
+      },
+      {
+        url: 'https://simple-twitter-api-2021.herokuapp.com',
+        description: 'Production server'
+      }
+    ]
+  },
+  apis: ['./routes/modules/*.js']
+}
 
 const app = express()
 const port = process.env.PORT || 3000
 
 app.use(bodyParser.urlencoded({ extended: true }))
 app.use(bodyParser.json())
+
+
+
+
+const swaggerDocs = swaggerJsDoc(swaggerOptions)
+app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocs))
+
 
 app.listen(port, () => console.log(`Example app listening on port ${port}!`))
